[WIP][ELF] Support LoongArch
Current status: confirmed working with things like binutils and QEMU;
can link LLVM but the produced llc/opt/etc. segfaults when doing
anything non-trivial.

Posting this for comments nevertheless, as I'm new to linker
development, and wanted to avoid making big mistakes early on (although
this part is debatable, as this is pretty complete by now).

Differential Revision: https://reviews.llvm.org/D138135