[LoongArch] Don't merge FrameIndex accesses into [F]{LD,ST}X
Otherwise eliminateFrameIndex cannot figure out how to fixup the stack
offset with its stateless logic, because there wouldn't be an immediate
slot for it to trivially write to, and it may not be easy to transform
the surrounding code to make it work.

This might not be the optimal fix but it fixes a fairly common crash
when compiling moderately complex code with Clang.

Differential Revision: https://reviews.llvm.org/D135251