[LoongArch] Support CodeModel::Large codegen
This is intended to behave like GCC's `-mcmodel=extreme`, and is a
prerequisite for enabling Clang-built Linux/LoongArch, as the percpu
variables apparently requires relocation of large offsets not
expressible with smaller code models currently implemented.

Technically the true GCC equivalent would be `-mcmodel=large` which is
not yet implemented there, and we probably do not want to take the
"Large" name until things settle in GCC side, but:

* LLVM does not have a `CodeModel::Extreme`, and it seems too early to
  have such a variant added just for enabling LoongArch;
* `CodeModel::Small` is already being used for GCC `-mcmodel=normal`
  which is already a case of divergent naming; and
* It is easy to migrate should additional CodeModel variants get added
  in the future,

So `CodeModel::Large` is synonymous with "extreme" for now.

Differential Revision: https://reviews.llvm.org/D150522